refactor: fold remaining nodejs_discovery files into runtime_detectors
Last of the Node.js special-casing in internal/nodemon: the detector functions
(isNodeProcess, resolveNodeVersion, release-URL regex) move into
runtime_detectors.go next to every other runtime's, resolveExePath moves to
proc_walk.go (it's a generic /proc helper), and maxNodeBinaryScanBytes becomes
maxBinaryScanBytes since it bounds all binary scans. Only jvm_* keeps dedicated
files — justified by its hsperfdata reader, flag parsing, and richer payload.

Also removes boolFromMap, orphaned by the NodeJs* snapshot-field removal
(caught by CI lint).
